PDFs can be created with Asciidoctor
The user can now choose in the settings if the PDF should be created with FOP or
Asciidoctor.
The goal was that the cheatsheet can not yet be rendered successfully.
The math and tree extension are not working.
Furthermore the uml block must be renamed to plantuml
The chart block is rendered via FxChartBlockProcessor.